refactor: Move `parseOnboardingURL` to `useAppBootstrap.functions.ts`
